editoast: rename module connection_pool to database
