+++ editoast/src/modelsv2/database.rs
@@ -3,11 +3,11 @@ use diesel_async::{
     AsyncPgConnection,
 };
 
-pub mod db_connection_error;
-pub mod db_connection_pool;
+pub mod connection_error;
+pub mod connection_pool;
 
-pub use db_connection_error::DbConnectionError;
-pub use db_connection_pool::DbConnectionPoolV2;
+pub use connection_error::DbConnectionError;
+pub use connection_pool::DbConnectionPoolV2;
 
 pub type DbConnection = AsyncPgConnection;
 pub type DbConnectionPool = Pool<DbConnection>;
