+++ editoast/src/modelsv2/database/connection_error.rs
